データ量の型をintからdoubleに変更

+++ src/mapred/org/apache/hadoop/mapred/IsolationRunner.java
@@ -58,7 +58,7 @@ public class IsolationRunner {
       LOG.info("Task " + taskid + " reporting done.");
     }
     
-    public void done(TaskAttemptID taskid, JvmContext jvmContext, int[] dataVolume)
+    public void done(TaskAttemptID taskid, JvmContext jvmContext, long[] dataVolume)
         throws IOException {
       LOG.info("Task " + taskid + " reporting done.");
     }
